python/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "MetaTag"
version = "1.3.0"
description = "MetaTag — Professional Audio Metadata Editor"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "RWAAHAHA", email = "3453-315h@users.noreply.github.com"}
]
keywords = ["audio", "metadata", "tagger", "audiobook", "mutagen", "pyside6"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "Topic :: Multimedia :: Sound/Audio :: Editors",
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: Microsoft :: Windows",
    "Operating System :: MacOS",
    "Operating System :: POSIX :: Linux",
]
dependencies = [
    "PySide6>=6.5.0",
    "mutagen>=1.47.0",
    "Pillow>=10.0.0",
]

[project.urls]
Homepage = "https://github.com/3453-315h/MetaTag"
Repository = "https://github.com/3453-315h/MetaTag"
"Bug Tracker" = "https://github.com/3453-315h/MetaTag/issues"

[project.optional-dependencies]
dev = ["pytest", "black", "flake8"]

[tool.setuptools.packages.find]
where = ["."]
include = ["metatag*"]

[project.scripts]
meta = "metatag.ui.main_window:main"
metatag = "metatag.ui.main_window:main"